Change behavior in update
The manual edit manager behavior has changed.
Now it waits for a user to pause typing for
X number of seconds before calling for an update.

Previous behavior was to call for an update once
X number of seconds had elapsed since the last
update.

The new behavior more reflects what we want, which
is we allow a user to make a burst of edits and then
when the user pauses, we show the result of those
edits.