cwl_dict has to be explicitly computed
This makes for more verbose code than the elegant
lazy computation previously implemented, however,
I want to be in control of this computation and
when it occurs. It's semi expensive and we anticipate
cases where bulk updates to many documents will occur.
I want this computation to only happen in the
update_from_code function in the editor window, so that
it happens JIT, and not accidentally in bulk, say. in
MultiDocumentManager.