wip: Saving state
A CwlView is a persistent object (whereas CwlDoc was
disposable). It takes over the functions of
DocumentManager. So I want to try out an approach
where we are allowed to create a CwlView only via
a special constructor, and subsequently, we can
only update it by changing the string and (if
a parent view) loading from disk.