Add ace-builds as a submodule
The plan is to keep it as a submodule so
we don't have 10 MB of files in our project.
For installing from git we will supply a post
install script (may not work on Windows) but
we will also package every thing neatly on Pypi
so windows users can pip install from that.