UI change: move command widget to editor pane
The command widget is now hideable with CMD+P.
It is part of a divider, so we can adjust how much
screen it takes when it is visible. So now, for
things that don't have a map we have no left pane,
but this we can replace by a command line preview
for CLT, javascript evaluation for expressions
etc.

This is a better use of space and paves the way for
our new idea of how to hook up workflow connections.