Implement jump to definition
This works for linked files in the run field and
in $import declarations.

This also implements the lookup table needed to
translate cursor position to document element.

I have chosen to add all the functions (completion,
jump to definition etc.) to the CWL type objects
themselves rather than create a separate class of
completer objects.