Move Any and linked process into specialized types
Instead of Any turning into an enum type we use a
special Any type.

Even though a linked process file behaves largely the
same as any other linked file, we subclass it because
we want to process the step interface while we validate
it.

We may decide in the end to pull this behavior out of
this class and do it all in one go with a different
data structure, but we will still need a distinct type
to make it easier to identify which are the steps with
linked processes.