Various fixes to allow key completions in empty LOM
This helps with key completions when we just start
with a LOM type (such as "in" or "requirements").
At this stage the LOM is just a string, so we assume
it's going to be the map form and just setup completions
as if that were the case.

This needs to be rewritten to look less chaotic and
ad hoc