Handle files linked via URL + show file contents on hover
A: Handle files linked via URL

While adding this feature I discovered that in all cases
we are loading the file contents as well (since we need
the contents to either parse the process interface, or
to load the Javascript or the user defined types).

Additionally, the most expensive part of the URL parsing
turns out to be establishing the TCP connection, it
costs about the same to also download the file contents.

For this reason, CWLLinkedFileType now caches the file
contents at the time of parsing.

B: show file contents on hover
B follows from A, being very easy to do as a result

Update tests.

Todo: Make schemadef loading use the cached contents

Closes #64
Closes #54