Load schemadefs during parsing + Show typedef on hover
We eliminate the repeated loading of the schema defs
by attaching the user defined types as we encounter
the linked type def files.

In the Record type class we now re-order the fields
such that if a "requirements" field exists it is
parsed first. This is necessary in order to load the
typedefs first so that they can be identified when
we parse inputs and outputs.

The typedef is now also shown on hover, which
properly closes #54