Show important step properties in graph
Information about scatter and conditionals in the
step are passed on to the workflow graph. The VS Code
plugin shows this information in the preview pane.
Scatter steps are marked with "S" and conditional
steps are marked with "?". Hovering over a step will
show the scatter variables and the conditional
expression.

Also refactor to rename "lines" variable to "line-numbers"
which is more clear in the context of a graph output.