Properly handle self-references in specification
In order to handle self-references in the specification
we need to place the still being constructed type in the
type dict before recursing into its children.

If we do not do this we see that the self-referential
child will remain a plain string in the types field
when we really need it to be one of the parent types.