Ensure type is properly initialized
Commit 8a8deee handles self-references in the
specification by placing the still being constructed
type in the type dict before recursing into its children.

This is the correct thing to do, but we forgot that now
the type information is incomplete when it is
initialized. Therefore the `required_fields` and
`all_fields` variables were not set correctly.

This caused type inference to fail in particular cases
such as for `Dirent`

We now re-run the init (the setup) once the type is
completely constructed ...

Version bumped to 2020.01.21. Happy New Year!

Closes #77
Closes #76